fix(consensus): use u128 validator stake throughout snapshots

+++ components/a-townchain/modules/atc-blockchain/kernel/src/consensus.rs
@@ -61,14 +61,14 @@ pub struct ConsensusEngine {
     pub proposer: String,
     height: Mutex<u64>,
     finalized: Mutex<Option<(u64, [u8; 32])>>,
-    slashed: Mutex<BTreeMap<String, u64>>,
+    slashed: Mutex<BTreeMap<String, u128>>,
     votes: Mutex<BTreeMap<[u8; 32], Vec<Vote>>>,
     validators: Mutex<BTreeMap<String, u64>>,
     validator_keys: Mutex<BTreeMap<String, [u8; 32]>>,
     /// Immutable validator-set snapshots keyed by the height at which the
     /// set became active. Consensus verification never falls back to the
     /// mutable current registry for historical blocks.
-    validator_snapshots: Mutex<BTreeMap<u64, (BTreeMap<String, u64>, BTreeMap<String, [u8; 32]>)>>,
+    validator_snapshots: Mutex<BTreeMap<u64, (BTreeMap<String, u128>, BTreeMap<String, [u8; 32]>)>>,
 }
 
 impl ConsensusEngine {
@@ -173,7 +173,7 @@ impl ConsensusEngine {
         self.validator_snapshots.lock().ok().map(|s| s.keys().copied().collect()).unwrap_or_default()
     }
 
-    pub fn register_validator(&self, address: String, stake: u64) -> Result<(), String> {
+    pub fn register_validator(&self, address: String, stake: u128) -> Result<(), String> {
         if address.is_empty() || stake == 0 {
             return Err("validator address and stake are required".into());
         }
@@ -256,7 +256,7 @@ impl ConsensusEngine {
         Ok(())
     }
 
-    pub fn validator_stake(&self, address: &str) -> u64 {
+    pub fn validator_stake(&self, address: &str) -> u128 {
         self.validators
             .lock()
             .unwrap()
